Fix minor text formatting bug
This commit fixes a minor error where a line that is exactly as long as
the output terminal would have a newline appended to it at the start of
the next line, resulting in an unsightly empty line if the colums were
set properly, or a too-long line if the game was set to use fewer cols
than the terminal window can display.

I think I've put more typing into this commit message than it took to
isolate and fix the bug.