Fix a minor equipment bug
Some equipment items had ID 0, and could not be removed if equipped.
This fix changes the "magic number" for no equipment to be -1 instead
of 0.

Also, the getEquipById() function now also requires the equipment type.
This is to ensure that it returns the proper item, since IDs are not
unique across different equipment types.